Refactor: Split paint methods from ShapeConverter into PaintConverter
Extract paint-related methods (fill, stroke, gradient, pattern) from
ShapeConverter into a new PaintConverter mixin for better separation
of concerns.

Changes:
- Create new PaintConverter mixin in src/psd2svg/core/paint.py
- Move 13 paint methods from ShapeConverter to PaintConverter
- Update Converter class to inherit from PaintConverter
- Add comprehensive module-level documentation to both mixins
- Clean up unused imports in shape.py

The ShapeConverter now focuses solely on geometric operations (paths,
shapes, boolean operations), while PaintConverter handles visual styling
(colors, gradients, patterns, strokes).

All tests pass. No functional changes.
